[MOV] base: move code linked to the users and groups UI
The code is moved to res_users.py and res_groups.py.  After refactoring
and introducing a new widget, the number of lines of that code is much
smaller and no longer deserves a separate file.
