+++ odoo/addons/base/__manifest__.py
@@ -68,6 +68,7 @@ The kernel of Odoo, needed for all installation.
         'views/res_bank_views.xml',
         'views/res_country_views.xml',
         'views/res_currency_views.xml',
+        'views/res_groups_views.xml',
         'views/res_users_views.xml',
         'views/res_device_views.xml',
         'views/res_users_identitycheck_views.xml',
